Move bitwise operations to a single file in compute
