--- src/compute/binary/mod.rs
@@ -1,2 +1,0 @@
-mod bitwise;
-pub use bitwise::*;
