+++ src/compute/bitwise.rs
