+++ src/compute/mod.rs
@@ -14,7 +14,7 @@
 pub mod aggregate;
 pub mod arithmetics;
 pub mod arity;
-pub mod binary;
+pub mod bitwise;
 pub mod boolean;
 pub mod boolean_kleene;
 pub mod cast;
