--- tests/it/compute/binary/mod.rs
@@ -1,1 +1,0 @@
-mod bitwise;
