+++ tests/it/compute/mod.rs
@@ -1,6 +1,6 @@
 mod aggregate;
 mod arithmetics;
-mod binary;
+mod bitwise;
 mod boolean;
 mod boolean_kleene;
 mod cast;
